requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
scipy>=1.12
statsmodels>=0.14
plotly>=5.22
kaleido>=0.2.1
openpyxl>=3.1
fpdf2>=2.7